refactor: remove 8.7 and 8.6 tests
