+++ qa/core-application-e2e-test-suite/.gitignore
@@ -6,4 +6,3 @@ node_modules/
 /playwright/.cache/
 .env
 .DS_Store
-utils/.auth
