--- qa/core-application-e2e-test-suite/c8_versions.json
@@ -1,5 +1,0 @@
-{
-  "8.8": "8.8",
-  "8.7": "8.7",
-  "8.6": "8.6"
-}
