+++ qa/core-application-e2e-test-suite/config/envs/.env.database.elasticsearch
@@ -1,0 +1,43 @@
+# Elasticsearch Configuration for Zeebe, Operate & Tasklist
+
+# Variables
+ZEEBE_CONTAINER_NAME=zeebe-${DATABASE}
+TASKLIST_CONTAINER_NAME=tasklist-${DATABASE}
+DATABASE_CONTAINER_NAME=${DATABASE}
+
+# Zeebe Configuration
+ZEEBE_BROKER_CLUSTER_PARTITIONS_COUNT=2
+ZEEBE_BROKER_EXPORTERS_CAMUNDAEXPORTER_CLASSNAME=io.camunda.exporter.CamundaExporter
+ZEEBE_BROKER_EXPORTERS_CAMUNDAEXPORTER_ARGS_CONNECT_URL=http://${DATABASE_CONTAINER_NAME}:9200
+ZEEBE_BROKER_EXPORTERS_CAMUNDAEXPORTER_ARGS_BULK_SIZE=100
+ZEEBE_BROKER_EXPORTERS_CAMUNDAEXPORTER_ARGS_INDEX_SHOULDWAITFORIMPORTERS=false
+ZEEBE_BROKER_GATEWAY_ENABLE=true
+ZEEBE_BROKER_NETWORK_HOST=zeebe
+ZEEBE_HOST=${ZEEBE_HOST:-}
+
+# Tasklist Configuration
+CAMUNDA_TASKLIST_ELASTICSEARCH_URL=http://${DATABASE_CONTAINER_NAME}:9200
+CAMUNDA_TASKLIST_ZEEBEELASTICSEARCH_URL=http://${DATABASE_CONTAINER_NAME}:9200
+CAMUNDA_TASKLIST_ZEEBE_GATEWAYADDRESS=${ZEEBE_CONTAINER_NAME}:26500
+CAMUNDA_TASKLIST_ZEEBE_RESTADDRESS=http://${ZEEBE_CONTAINER_NAME}:8080
+CAMUNDA_TASKLIST_BACKUP_REPOSITORYNAME=test
+CAMUNDA_TASKLIST_DATABASE=${DATABASE}
+
+# Operate Configuration
+CAMUNDA_OPERATE_TASKLIST_URL=http://localhost:8080
+CAMUNDA_OPERATE_ZEEBE_GATEWAYADDRESS=${ZEEBE_CONTAINER_NAME}:26500
+CAMUNDA_OPERATE_ZEEBE_RESTADDRESS=http://${ZEEBE_CONTAINER_NAME}:8080
+CAMUNDA_OPERATE_ELASTICSEARCH_URL=http://${DATABASE_CONTAINER_NAME}:9200
+CAMUNDA_OPERATE_ZEEBEELASTICSEARCH_URL=http://${DATABASE_CONTAINER_NAME}:9200
+CAMUNDA_OPERATE_BACKUP_REPOSITORYNAME=test
+CAMUNDA_OPERATE_DATABASE=${DATABASE}
+
+# General Configuration
+SPRINGFOX_DOCUMENTATION_SWAGGER_V2_PATH=/documentation
+CAMUNDA_REST_QUERY_ENABLED=true
+CAMUNDA_DATABASE_URL=http://${DATABASE_CONTAINER_NAME}:9200
+
+# External URLs
+EXT_CAMUNDA_TASKLIST_ZEEBE_GATEWAYADDRESS=localhost:26500
+EXT_CAMUNDA_TASKLIST_ZEEBE_RESTADDRESS=http://localhost:8089
+EXT_ELASTICSEARCH_URL=http://localhost:9200
