--- qa/core-application-e2e-test-suite/pages/8.7/OperateProcessInstancePage.ts
@@ -1,86 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-import {Page, Locator, expect} from '@playwright/test';
-import {sleep} from 'utils/sleep';
-
-class OperateProcessInstancePage {
-  private page: Page;
-  readonly diagram: Locator;
-  readonly completedIcon: Locator;
-  readonly diagramSpinner: Locator;
-  readonly activeIcon: Locator;
-  readonly variablesList: Locator;
-  readonly messageVariable: Locator;
-  readonly statusVariable: Locator;
-
-  constructor(page: Page) {
-    this.page = page;
-    this.diagram = page.getByTestId('diagram');
-    this.completedIcon = page
-      .getByTestId('instance-header')
-      .getByTestId('COMPLETED-icon');
-    this.diagramSpinner = page.getByTestId('diagram-spinner');
-    this.activeIcon = page
-      .getByTestId('instance-header')
-      .getByTestId('ACTIVE-icon');
-    this.variablesList = page.getByTestId('variables-list');
-    this.messageVariable = page.getByTestId('variable-message');
-    this.statusVariable = page.getByTestId('variable-status');
-  }
-
-  async connectorResultVariableName(name: string): Promise<Locator> {
-    return this.page.getByTestId(name);
-  }
-
-  async connectorResultVariableValue(variableName: string): Promise<Locator> {
-    return this.page.getByTestId(variableName).locator('td').last();
-  }
-
-  async completedIconAssertion(): Promise<void> {
-    let retryCount = 0;
-    const maxRetries = 3;
-    while (retryCount < maxRetries) {
-      try {
-        await expect(this.completedIcon).toBeVisible({
-          timeout: 90000,
-        });
-        return; // Exit the function if the expectation is met
-      } catch {
-        // If the active icon isn't found, reload the page and try again
-        retryCount++;
-        console.log(`Attempt ${retryCount} failed. Retrying...`);
-        await this.page.reload();
-        await sleep(10000);
-      }
-    }
-    throw new Error(`Active icon not visible after ${maxRetries} attempts.`);
-  }
-
-  async activeIconAssertion(): Promise<void> {
-    let retryCount = 0;
-    const maxRetries = 3;
-    while (retryCount < maxRetries) {
-      try {
-        await expect(this.activeIcon).toBeVisible({
-          timeout: 90000,
-        });
-        return; // Exit the function if the expectation is met
-      } catch {
-        // If the active icon isn't found, reload the page and try again
-        retryCount++;
-        console.log(`Attempt ${retryCount} failed. Retrying...`);
-        await this.page.reload();
-        await sleep(10000);
-      }
-    }
-    throw new Error(`Active icon not visible after ${maxRetries} attempts.`);
-  }
-}
-
-export {OperateProcessInstancePage};
