--- qa/core-application-e2e-test-suite/pages/8.7/OperateProcessesPage.ts
@@ -1,93 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-import {Page, Locator} from '@playwright/test';
-import {sleep} from 'utils/sleep';
-
-class OperateProcessesPage {
-  private page: Page;
-  readonly processResultCount: Locator;
-  readonly processActiveCheckbox: Locator;
-  readonly processCompletedCheckbox: Locator;
-  readonly processRunningInstancesCheckbox: Locator;
-  readonly processIncidentsCheckbox: Locator;
-  readonly processPageHeading: Locator;
-  readonly noMatchingInstancesMessage: Locator;
-  readonly processFinishedInstancesCheckbox: Locator;
-
-  constructor(page: Page) {
-    this.page = page;
-    this.processResultCount = page.getByTestId('result-count');
-    this.processActiveCheckbox = page
-      .locator('label')
-      .filter({hasText: 'Active'});
-    this.processCompletedCheckbox = page
-      .locator('label')
-      .filter({hasText: 'Completed'});
-    this.processRunningInstancesCheckbox = page
-      .locator('label')
-      .filter({hasText: 'Running Instances'});
-    this.processIncidentsCheckbox = page
-      .locator('label')
-      .filter({hasText: 'Incidents'});
-    this.processPageHeading = page
-      .getByTestId('expanded-panel')
-      .getByRole('heading', {name: 'Process'});
-    this.noMatchingInstancesMessage = page.getByText(
-      'There are no Instances matching this filter set',
-    );
-    this.processFinishedInstancesCheckbox = page
-      .getByTestId('filter-finished-instances')
-      .getByRole('checkbox');
-  }
-
-  async clickProcessActiveCheckbox(): Promise<void> {
-    await this.processActiveCheckbox.click();
-  }
-
-  async clickProcessCompletedCheckbox(): Promise<void> {
-    await this.processCompletedCheckbox.click({timeout: 120000});
-  }
-
-  async clickProcessIncidentsCheckbox(): Promise<void> {
-    await this.processIncidentsCheckbox.click({timeout: 90000});
-  }
-
-  async clickRunningProcessInstancesCheckbox(): Promise<void> {
-    await this.processRunningInstancesCheckbox.click({timeout: 90000});
-  }
-
-  async clickFinishedProcessInstancesCheckbox(): Promise<void> {
-    await this.processFinishedInstancesCheckbox.click({timeout: 90000});
-  }
-
-  async clickProcessInstanceLink(processName: string): Promise<void> {
-    let retryCount = 0;
-    const maxRetries = 3;
-    while (retryCount < maxRetries) {
-      try {
-        await this.page
-          .locator(`td:right-of(:text("${processName}"))`)
-          .first()
-          .click({timeout: 90000});
-        return; // Exit the function if the click is successful
-      } catch {
-        // If process isn't found, reload the page and try again
-        retryCount++;
-        console.log(`Attempt ${retryCount} failed. Retrying...`);
-        await this.page.reload();
-        await sleep(10000);
-      }
-    }
-    throw new Error(
-      `Failed to click on process instance link after ${maxRetries} attempts.`,
-    );
-  }
-}
-
-export {OperateProcessesPage};
