+++ qa/core-application-e2e-test-suite/pages/OperateHomePage.ts
