+++ qa/core-application-e2e-test-suite/tests/hto-user-flows.spec.ts
@@ -7,20 +7,22 @@
  */
 
 import {expect} from '@playwright/test';
-import {test} from '@fixtures/8.7';
+import {test} from 'fixtures';
 import {createInstances, deploy} from 'utils/zeebeClient';
-import {navigateToApp} from '@pages/8.7/UtilitiesPage';
+import {navigateToApp} from '@pages/UtilitiesPage';
 import {sleep} from 'utils/sleep';
 import {captureScreenshot, captureFailureVideo} from '@setup';
 
 test.beforeAll(async () => {
-  await deploy(['./resources/Job_Worker_Process.bpmn',
-  './resources/Variable_Process.bpmn',
-  './resources/Zeebe_User_Task_Process.bpmn',
-  './resources/New Form.form',
-  './resources/User_Task_Process_With_Form.bpmn',
-  './resources/Start_Form_Process.bpmn',
-  './resources/Zeebe_User_Task_Process_With_Priority.bpmn']);
+  await deploy([
+    './resources/Job_Worker_Process.bpmn',
+    './resources/Variable_Process.bpmn',
+    './resources/Zeebe_User_Task_Process.bpmn',
+    './resources/New Form.form',
+    './resources/User_Task_Process_With_Form.bpmn',
+    './resources/Start_Form_Process.bpmn',
+    './resources/Zeebe_User_Task_Process_With_Priority.bpmn',
+  ]);
   await createInstances('Job_Worker_Process', 1, 1);
   await createInstances('Zeebe_User_Task_Process', 1, 1);
   await createInstances('Variable_Process', 1, 1, {
@@ -35,7 +37,7 @@ test.describe('HTO User Flow Tests', () => {
   test.beforeEach(async ({page, taskListLoginPage, taskPanelPage}) => {
     await navigateToApp(page, 'tasklist');
     await taskListLoginPage.login('demo', 'demo');
-    await taskPanelPage.taskListBannerIsVisible();
+    await expect(taskPanelPage.taskListPageBanner).toBeVisible();
   });
 
   test.afterEach(async ({page}, testInfo) => {
@@ -63,7 +65,7 @@ test.describe('HTO User Flow Tests', () => {
     await test.step('View Process Instance in Operate, complete User Task in Tasklist & assert process complete in Operate', async () => {
       await navigateToApp(page, 'operate');
       await operateLoginPage.login('demo', 'demo');
-      await operateHomePage.operateBannerIsVisible();
+      await expect(operateHomePage.operateBanner).toBeVisible();
       await expect(operateHomePage.processesTab).toBeVisible({timeout: 180000});
       await operateHomePage.clickProcessesTab();
       await operateProcessesPage.clickProcessInstanceLink('Job_Worker_Process');
