--- qa/core-application-e2e-test-suite/utils/.auth
@@ -1,1 +1,0 @@
-{}
